requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
requests